// src/LupiraCalWeb.Client/src/data/api/models/placeKind.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Lupira Cal API
 * Calendar, contacts, and CalDAV backend for Lupira. Authenticate with a Bearer token issued by the OIDC provider (Authentik).
 * OpenAPI spec version: v1
 */

export type PlaceKind = typeof PlaceKind[keyof typeof PlaceKind];


export const PlaceKind = {
  Country: 'Country',
  City: 'City',
  Address: 'Address',
  Venue: 'Venue',
} as const;